requirements.txt
flask>=2.3.0
nltk>=3.8.1
scikit-learn>=1.3.0
numpy>=1.24.0
pandas>=2.0.0
python-docx>=0.8.11
PyPDF2>=3.0.0
rank-bm25>=0.2.2
colorama>=0.4.6
tabulate>=0.9.0
